Break action into Action and MultiAction components.
