+++ src/components/MultiAction/index.ts
@@ -1,0 +1,4 @@
+export { default } from './MultiAction'
+
+// NOTE: uncomment this line to pass through named exports
+// export * from './MultiAction'
